requirements.txt
blessed==1.9.5
botocore==1.4.35
cement==2.8.2
colorama==0.3.7
Django==1.9.7
django-filter==0.13.0
django-rest-swagger==0.3.8
djangorestframework==3.3.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
httpie==0.9.4
jmespath==0.9.0
Markdown==2.6.6
MySQL-python==1.2.5
pathspec==0.3.4
Pygments==2.1.3
python-dateutil==2.5.3
PyYAML==3.11
requests==2.9.1
semantic-version==2.5.0
six==1.10.0
texttable==0.8.4
wcwidth==0.1.7
websocket-client==0.37.0
